% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mammalcol-package.R
\docType{package}
\name{mammalcol-package}
\alias{mammalcol}
\alias{mammalcol-package}
\title{mammalcol: Access to the List of Mammal Species of Colombia}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'mammalcol' is to provide easy access to a meticulously structured dataset of Colombian mammal species in R. The 2026 update includes comprehensive, detailed species accounts, and distribution information at departamental and municipal level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dlizcano/mammalcol}
  \item \url{https://dlizcano.github.io/mammalcol/}
  \item Report bugs at \url{https://github.com/dlizcano/mammalcol/issues}
}

}
\author{
\strong{Maintainer}: Diego J. Lizcano \email{dj.lizcano@gmail.com} (\href{https://orcid.org/0000-0002-9648-0576}{ORCID})

Authors:
\itemize{
  \item Cristian A. Cruz-Rodríguez \email{cristian.alexander.cruz.rodriguez@umontreal.ca} (\href{https://orcid.org/0000-0002-8319-4706}{ORCID}) [contributor]
  \item Andres F. Suárez-Castro \email{a.suarezcastro@griffith.edu.au} (\href{https://orcid.org/0000-0002-6621-3821}{ORCID}) [contributor]
  \item Baltazar González \email{baltazargch@gmail.com} (\href{https://orcid.org/0000-0001-5805-8107}{ORCID}) [contributor]
  \item Karol Aguilar-Rios \email{Kaguilar1d2002@gmail.com} (\href{https://orcid.org/0009-0002-6331-2653}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Alejandra Bonilla-Sánchez \email{alejabonilla.sanchez@gmail.com} (\href{https://orcid.org/0000-0003-0803-5884}{ORCID}) [contributor]
  \item Camilo A. Calderón-Acevedo \email{calderon.mammalogy@gmail.com} (\href{https://orcid.org/0000-0002-1468-3565}{ORCID}) [contributor]
}

}
\keyword{internal}
