% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mammalmap.R
\name{mammalmap}
\alias{mammalmap}
\title{Plot a map Data for one species of the List of the Mammals of Colombia}
\usage{
mammalmap(species)
}
\arguments{
\item{species}{A character string containing the name of the species to plot the map.}
}
\value{
A map for the species.
}
\description{
This function retrieve the locality data for one species in
the MammalCol pacakage dataset, and returns a map of Departamentos
where the species has been recorded.
}
\details{
The function does not allows fuzzy matching, so the specie name should be correct.
It is advised to run first the search_mammalcol function.
}
\examples{

species <- "Tapirus pinchaque" 
mammalmap(species)

}
