% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_coordiantes.R
\docType{data}
\name{test_data_coordiantes}
\alias{test_data_coordiantes}
\title{test_data_coordiantes Dataset}
\format{
A tibble with 548 rows and 19 columns:
\describe{
\item{id}{id from original taxon table.}
}
}
\usage{
test_data_coordiantes
}
\description{
In progress
}
\details{
XXXXX
}
\examples{

# Load the mammalcol package
library(mammalcol)

# Access the mammalcol_tab dataset
data("test_data_coordiantes")

# Display the first few rows
head(test_data_coordiantes)


}
\references{
The dataset is based on the "List of the Mammals of Colombia" by:
Ramírez-Chaves H E, Leuro Robles N G, Castaño Rivera A, Morales-Martínez D M,
Suárez Castro A F, Rodríguez-Posada M E, Zurc D, Concha Osbahr D C, Trujillo A,
Noguera Urbano E A, Pantoja Peña G E, González Maya J F, Pérez Torres J,
Mantilla Meluk H, López Castañeda C, Velásquez Valencia A, Zárrate Charry D (2024):
Mamíferos de Colombia. v1.14. Sociedad Colombiana de Mastozoología. Dataset/Checklist
\href{https://ipt.biodiversidad.co/sib/resource?r=mamiferos_col#anchor-citation}{SiB Colombia}
}
\seealso{
For more information about the "mammalcol" package and the data sources, visit
the package's GitHub repository: \url{https://github.com/dlizcano/mammalcol}
}
\author{
Data compilation: Ramírez-Chaves 2021, Package implementation: Cristian A. Cruz-R.
}
\keyword{datasets}
